requirements.txt
moviepy==1.0.3
Pillow==9.5.0
aiohttp==3.11.10
aiofiles==24.1.0
numpy